+++ rth/adapters/general_adapter.py
@@ -5,9 +5,10 @@ from rth.core.errors import WrongOptionName
 class GeneralAdapter:
     file_path = None
 
-    def __init__(self, file_path, ignore=False):
+    def __init__(self, file_path, ignore=False, debug=False):
         self.file_path = file_path
         self.ignore = ignore
+        self.debug = debug
 
     def __call__(self, *args, **kwargs):
         return self.evaluate()
@@ -16,7 +17,7 @@ class GeneralAdapter:
         pass
 
     def _process(self, subnetworks, routers, links, options=None):
-        dispatch = Dispatcher()
+        dispatch = Dispatcher(debug=self.debug)
 
         if options:
             for name in options: