+++ CHANGELOG.md
@@ -7,3 +7,4 @@
   - (dev side) added a general adapter to ease implementation of new formats in the future
   - fixed bugs concerning the master router and route (See [issue #6](https://github.com/BioTheWolff/RTH/issues/6))
   - reworked the way the master route is handled and processed in the RTG class
+  - added an option for preferred routers